expanded number range for math.gcd and math.lcm

+++ include/gura/Algorithm.h
@@ -18,21 +18,21 @@ inline double DegToRad(double n) { return Math_PI * n / 180; }
 //-----------------------------------------------------------------------------
 // GCD/LCM
 //-----------------------------------------------------------------------------
-GURA_DLLDECLARE int CalcGCD(int a, int b);
-GURA_DLLDECLARE int CalcLCM(int a, int b);
+GURA_DLLDECLARE Int64 CalcGCD(Int64 a, Int64 b);
+GURA_DLLDECLARE Int64 CalcLCM(Int64 a, Int64 b);
 
 //-----------------------------------------------------------------------------
 // CRC32
 //-----------------------------------------------------------------------------
 class GURA_DLLDECLARE CRC32 {
 private:
-	ULong _crc32;
+	UInt32 _crc32;
 	size_t _bytes;
 public:
 	inline CRC32() : _crc32(0xffffffff), _bytes(0) {}
 	inline void Initialize() { _crc32 = 0xffffffff, _bytes = 0; }
 	void Update(const void *buff, size_t bytes);
-	inline ULong GetResult() const { return ~_crc32 & 0xffffffff; }
+	inline UInt32 GetResult() const { return ~_crc32 & 0xffffffff; }
 	inline size_t GetBytes() const { return _bytes; }
 };
